requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-filter==23.4
idna==3.4
lxml==4.9.3
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
